chore: cleanup code,  place create Team functions inside of createteamform
